requirements.txt
python-telegram-bot>=20,<22
requests>=2.31.0
